Added a new MonadK primitive
Turns out we needed a bit more complexity to achieve certain buttons.
Now you can throw events into the front of KMonad using the `inject`
action. This seems like it comes with a lot of power, so it should be
handled carefully. At the moment we are only using this to rethrow
certain events.