Additional release command for `cmd-button'
With this change, as well as the help of the `around' function, one is
able to use a `cmd-button' for e.g. switching the keyboard layout (via
`setxkbmap' or a similar utility) after toggling a layer and then
switching back to the normal layout after the toggle has been released.