nixos-module: Restart on failure
Automatically restart the service upon failure, emulating
functionality that the normal SystemD service already has.  This
ensures that all connected keyboards are initialized by the system.

A long-term plan would be to integrate the service more closely with
the NixOS module, such that these kinds of issues can be fixed
directly within the service, without touching the Nix functionality.

Fixes: https://github.com/kmonad/kmonad/issues/292