Correcting Windows virtual-key code translation
Kmonad uses a map to associate Windows virtual-key codes with KMonad
KeyCodes. When translating the other direction, the logic would flip the
entries of this map and go from KeyCode to virtual-key code. However,
where more than one virtual-key code mapped to the same KMonad KeyCode,
this created a conflict where only one would be able to be sent to the
sink, and often not the expected code.

This commit fixes that issue by maintaining two separate maps, one for
each direction. In this way, multiple virtual-key codes may be
interpreted as the same KMonad KeyCode, and multiple KeyCodes may be
sent as the same virtual-key code as necessary.

See https://github.com/kmonad/kmonad/issues/326