optional :timeout-button for tap-hold-next-release
If the :timeout-button argument is specified, tap-hold-next-release will
switch to that button instead of the hold button if the timeout elapses.
This makes it possible to make implement a button which can both be
tapped and pressed-and-held in order to type 1 or multiple (via the
operating system's auto-repeat functionality) copies of a character,
while still behaving as a modifier key when pressed in combination with
other keys.